+++ dev/src/pattern_lexer.js
@@ -108,7 +108,7 @@
             return captureVals(TOKENS.OP.rgx, pattern);
         }
 
-        function compilePattern(pattern) {
+        function compilePattern(pattern, ignoreCase) {
             pattern = pattern || '';
 
             if(pattern){
@@ -135,7 +135,7 @@
                 //single slash is treated as empty and end slash is optional
                 pattern += '\\/?';
             }
-            return new RegExp('^'+ pattern + '$');
+            return new RegExp('^'+ pattern + '$', ignoreCase? 'i' : '');
         }
 
         function replaceTokens(pattern, regexpName, replaceName) {