+++ dev/tests/spec/parse.spec.js
@@ -7,14 +7,24 @@ var crossroads = crossroads || require('crossroads');
 
 describe('crossroads.parse()', function(){
 
+    var _prevTypecast;
+
+
+    beforeEach(function(){
+        _prevTypecast = crossroads.shouldTypecast;
+    });
+
 
     afterEach(function(){
         crossroads.resetState();
         crossroads.removeAllRoutes();
         crossroads.routed.removeAll();
         crossroads.bypassed.removeAll();
+        crossroads.shouldTypecast = _prevTypecast;
     });
 
+
+    // ---
 
 
     describe('simple string route', function(){
@@ -233,7 +243,6 @@ describe('crossroads.parse()', function(){
     describe('typecast values', function(){
 
         it('should typecast values if shouldTypecast is set to true', function(){
-            var prevTypecast = crossroads.shouldTypecast;
             crossroads.shouldTypecast = true;
 
             var t1, t2, t3, t4, t5, t6;
@@ -256,12 +265,9 @@ describe('crossroads.parse()', function(){
             expect( t4 ).toBe( false );
             expect( t5 ).toBe( null );
             expect( t6 ).toBe( undefined );
-
-            crossroads.shouldTypecast = prevTypecast; //restore
         });
 
         it('should not typecast if shouldTypecast is set to false', function(){
-            var prevTypecast = crossroads.shouldTypecast;
             crossroads.shouldTypecast = false;
 
             var t1, t2, t3, t4;
@@ -280,8 +286,6 @@ describe('crossroads.parse()', function(){
             expect( t2 ).toBe( '123' );
             expect( t3 ).toBe( 'true' );
             expect( t4 ).toBe( 'false' );
-
-            crossroads.shouldTypecast = prevTypecast; //restore
         });
 
     });
@@ -906,6 +910,8 @@ describe('crossroads.parse()', function(){
 
         describe('required query string after required segment', function () {
             it('should parse query string into an object and typecast vals', function () {
+                crossroads.shouldTypecast = true;
+
                 var r = crossroads.addRoute('{a}{?b}');
                 var t1, t2;
                 r.matched.addOnce(function(a, b){
@@ -921,6 +927,8 @@ describe('crossroads.parse()', function(){
 
         describe('required query string after optional segment', function () {
             it('should parse query string into an object and typecast vals', function () {
+                crossroads.shouldTypecast = true;
+
                 var r = crossroads.addRoute(':a:{?b}');
                 var t1, t2;
                 r.matched.addOnce(function(a, b){
@@ -946,6 +954,8 @@ describe('crossroads.parse()', function(){
 
         describe('optional query string after required segment', function () {
             it('should parse query string into an object and typecast vals', function () {
+                crossroads.shouldTypecast = true;
+
                 var r = crossroads.addRoute('{a}:?b:');
                 var t1, t2;
                 r.matched.addOnce(function(a, b){
@@ -971,6 +981,8 @@ describe('crossroads.parse()', function(){
 
         describe('optional query string after optional segment', function () {
             it('should parse query string into an object and typecast vals', function () {
+                crossroads.shouldTypecast = true;
+
                 var r = crossroads.addRoute(':a::?b:');
                 var t1, t2;
                 r.matched.addOnce(function(a, b){
@@ -994,6 +1006,21 @@ describe('crossroads.parse()', function(){
             });
         });
 
+        describe('optional query string after required segment without typecasting', function () {
+            it('should parse query string into an object and not typecast vals', function () {
+                var r = crossroads.addRoute('{a}:?b:');
+                var t1, t2;
+
+                r.matched.addOnce(function(a, b){
+                    t1 = a;
+                    t2 = b;
+                });
+                crossroads.parse('foo.php?lorem=ipsum&asd=123&bar=false');
+
+                expect( t1 ).toEqual( 'foo.php' );
+                expect( t2 ).toEqual( {lorem : 'ipsum', asd : '123', bar : 'false'} );
+            });
+        });
     });