+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Crossroads.js Changelog #
 
+## Next ##
+
+### API Changes
+
+ - add `crossroads.pipe()` and `crossroads.unpipe()` (#70)
+ - added way to toggle case sensitivity `crossroads.ignoreCase`, default is
+   `true` (#53)
+ - add `crossroads.ignoreState`. (#57)
+
+### Improvements
+
+ - `decodeQueryString()` now respects `shouldTypecast` (#71)
+ - changed `Route.rules` array validation to be case insensitive if
+   `crossroads.ignoreCase = true` (#49)
+
+
+
 ## v0.10.0 (2012/08/12) ##
 
 ### Improvements